requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
SQLAlchemy==2.0.36
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==43.0.3
python-multipart==0.0.9
pydantic==2.9.2